Agent/embedding/requirements.txt
# --- START OF FILE requirements.txt ---
# Core Dependencies
pymongo>=4.5.0
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
nltk>=3.8.1

# For API server
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=1.10.5
httpx>=0.23.0

# For testing (optional but recommended)
pytest>=7.3.1
pytest-asyncio>=0.21.0